src/main/webapp/app/entities/syllabus/route/syllabus-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ISyllabus, Syllabus } from '../syllabus.model';
import { SyllabusService } from '../service/syllabus.service';

import { SyllabusRoutingResolveService } from './syllabus-routing-resolve.service';

describe('Syllabus routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: SyllabusRoutingResolveService;
  let service: SyllabusService;
  let resultSyllabus: ISyllabus | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(SyllabusRoutingResolveService);
    service = TestBed.inject(SyllabusService);
    resultSyllabus = undefined;
  });

  describe('resolve', () => {
    it('should return ISyllabus returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSyllabus = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSyllabus).toEqual({ id: 123 });
    });

    it('should return new ISyllabus if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSyllabus = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultSyllabus).toEqual(new Syllabus());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Syllabus })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSyllabus = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSyllabus).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
